chore: update DistributionScheduler deployment to new contract
