+++ pkg/deployments/action-ids/goerli/action-ids.json
@@ -372,16 +372,6 @@
       }
     }
   },
-  "20220422-distribution-scheduler": {
-    "DistributionScheduler": {
-      "useAdaptor": true,
-      "actionIds": {
-        "scheduleDistribution(address,address,uint256,uint256)": "0x73d1abbe8fdc1eada717f73299f34b7217a71930114175a6cede25a4735d27b6",
-        "startDistributionForToken(address,address)": "0x5b56750a9db9bac83e09e56bb50b0420b86c5bb5255075f7d616ce3700ef771a",
-        "startDistributions(address)": "0xc0b53b0850f307966f922a3dd4dfa9956a28f9db59b76735e38aacf43b63b35b"
-      }
-    }
-  },
   "20220425-unbutton-aave-linear-pool": {
     "UnbuttonAaveLinearPoolFactory": {
       "useAdaptor": true,
