--- pkg/deployments/tasks/20220422-distribution-scheduler/output/goerli.json
@@ -1,4 +1,0 @@
-{
-  "DistributionScheduler": "0xbD84e83b164aaCa0E2DABf3Aa27a9B7A1d62d200",
-  "timestamp": 1656957523632
-}
