--- pkg/deployments/tasks/20220422-distribution-scheduler/readme.md
@@ -1,8 +1,0 @@
-# 2022-04-22 - Distribution Scheduler
-
-Deployment of the `DistributionScheduler`, which can be used to schedule token distributions to gauges in advance, making operations simpler.
-
-## Useful Files
-
-- [Ethereum mainnet addresses](./output/mainnet.json)
-- [`DistributionScheduler` ABI](./abi/DistributionScheduler.json)
